Extract AdditionalPrinterColumn.Format enum to PrinterColumnFormat
